Added nav, model, primitive data binding
Still no styling. Also added a global event listener that uses the backbone history functionality to change pages.

I've used a single model for both views (as I think at this point it's all I require). There's the beginnings of the databinding being fleshed out, but it's only the beginning